requirements.txt
torch==2.0.1+cu118
torchvision==0.15.2+cu118
torch-geometric==2.3.1
torch-scatter==2.1.1+pt20cu118
torch-sparse==0.6.17+pt20cu118
torch-cluster==1.6.1+pt20cu118
torch-spline-conv==1.2.2+pt20cu118
openslide-python==1.2.0
numpy==1.24.3
pandas==2.0.2
scikit-learn==1.2.2
tqdm==4.65.0
